feat(frontend): separate imports page
